fix(core): Path with non-existent sub directories fails to be made
Refers to the "makeDirectory" method. It used to create a single
directory just fine, however it would refuse to create a sub
directory if the parent did not exist. This is a necessary
function in nutra for handling temporary folders. With the
help of the "mkdirp" npm module, it now creates the
directories recursively.